REST API: adds revision attribute to /sites/$site/posts/$post endpoint
This will be used in calypso to show inside the editor revisions that have been saved for a given Post and to link to the latest one. This attribute is only added when the Post is requested with ?conext=edit, so as not to slow down our many other posts-list queries.

Merges r126050-wpcom.