Photon: Fall back to returning no sizes in photon image if no size meta
We were trying to give the $image a size no matter what, which was causing problems when wp_get_attachment_metadata() was failing to give us the meta for whatever reason.

Instead of fighting it, let's just fall back to not returning sizes with the image.  The sacrifice is that this image won't have srcset in WP>4.4